# gpt-config/openapi-actions.yaml
openapi: "3.1.0"
info:
  title: Course Companion FTE – Actions API
  description: >
    FastAPI backend for the Course Companion FTE ChatGPT Custom GPT.
    Provides course content, quiz submission, progress tracking, and search.
    All endpoints require a Bearer JWT token (obtained from /auth/login).
    Zero LLM calls happen on the backend — all intelligence is in this GPT.
  version: "1.0.0"
servers:
  - url: https://your-backend-domain.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChapterSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
          description: Chapter number (1-5)
        slug:
          type: string
          example: "what-is-generative-ai"
        title:
          type: string
        summary:
          type: string
          description: One-sentence chapter description
        estimated_mins:
          type: integer
          description: Estimated reading time in minutes
        tier_required:
          type: string
          enum: [free, pro]
        has_quiz:
          type: boolean
        tags:
          type: array
          items:
            type: string
    ChapterDetail:
      allOf:
        - $ref: '#/components/schemas/ChapterSummary'
        - type: object
          properties:
            content_md:
              type: string
              description: Full chapter content in Markdown format
...
    QuizAttemptRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
          description: Map of question_id → selected answer (MCQ) or text answer (open)
          example:
            "q1": "B"
            "q2": "Transformer architecture"
    QuizAttemptResult:
      type: object
      properties:
        attempt_id:
          type: string
          format: uuid
        score:
          type: number
          description: Score as decimal (0.0 – 1.0)
        score_percent:
          type: integer
          description: Score as percentage (0-100)
        passed:
          type: boolean
        correct_count:
          type: integer
        total_questions:
          type: integer
        feedback:
          type: array
          items:
            type: object
            properties:
              question_id:
                type: string
              correct:
                type: boolean
              correct_answer:
                type: string
              explanation:
                type: string
        attempts_today:
          type: integer
          description: Number of attempts made today (free tier limit is 3)
        attempts_remaining:
          type: integer
          description: Remaining attempts today (-1 means unlimited for Pro)
...
paths:
...
  /chapters/{chapter_id}:
    get:
      operationId: getChapter
      summary: Get full chapter content
      description: >
        Returns the complete chapter including full markdown content.
        Free users can only access chapters 1-3. Returns 403 for locked chapters.
        Call this when the user wants to read a specific chapter.
      tags: [Chapters]
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
          description: Chapter UUID or slug (e.g. "what-is-generative-ai")
      responses:
        "200":
          description: Full chapter detail with content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterDetail'
        "403":
          description: Pro tier required
        "404":
          description: Chapter not found
...
  /quizzes/{quiz_id}/attempt:
    post:
      operationId: submitQuizAttempt
      summary: Submit quiz answers and get results
      description: >
        Submit answers for a quiz and receive a scored result with feedback.
        Grading is entirely rule-based — no LLM required.
        Free users limited to 3 attempts per quiz per day.
        Returns per-question feedback with correct answers and explanations.
      tags: [Quizzes]
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizAttemptRequest'
      responses:
        "200":
          description: Quiz results with per-question feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizAttemptResult'
        "429":
          description: Daily attempt limit reached (free tier)
  /quizzes/{quiz_id}/attempts:
    get:
      operationId: getQuizAttempts
      summary: Get attempt history for a quiz
      description: Returns the user's previous attempts for a specific quiz, ordered by most recent first.
      tags: [Quizzes]
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of previous attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    attempt_id:
                      type: string
                    score:
                      type: number
                    score_percent:
                      type: integer
                    passed:
                      type: boolean
                    attempted_at:
                      type: string
                      format: date-time

  # ── Progress ───────────────────────────────────────────────────────────────
...
  /progress/chapters/{chapter_id}:
    put:
      operationId: updateChapterProgress
      summary: Update chapter reading status
      description: >
        Mark a chapter as in_progress (when user starts reading) or completed (when done).
        Use status="in_progress" when user opens a chapter, "completed" when they finish.
      tags: [Progress]
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [not_started, in_progress, completed]
      responses:
        "200":
          description: Progress updated
  /progress/analytics:
    get:
      operationId: getProgressAnalytics
      summary: Get detailed progress analytics (Pro only)
      description: Returns detailed analytics including average quiz scores, completion streak, and time spent. Pro tier required.
      tags: [Progress]
      responses:
        "200":
          description: Detailed analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  avg_quiz_score:
                    type: number
                  best_quiz_score:
                    type: number
                  quizzes_passed:
                    type: integer
                  quizzes_failed:
                    type: integer
                  chapters_completed:
                    type: integer
                  total_time_mins:
                    type: integer
                  current_streak_days:
                    type: integer
        "403":
          description: Pro tier required

  # ── Search ─────────────────────────────────────────────────────────────────
...